Fixed errors in Venda.cpp
