+++ src/main.cpp
@@ -4,9 +4,7 @@
 
 #include "Cliente.hpp"
 #include "Funcionario.hpp"
-#include "Especialista.hpp"
 #include "Venda.hpp"
-#include "Gerente.hpp"
 
 int main()
 {
